rename iql prompt format
