+++ src/dbally/gradio/gradio_interface.py
@@ -9,7 +9,7 @@ from dbally import BaseStructuredView
 from dbally.audit import CLIEventHandler
 from dbally.collection import Collection
 from dbally.collection.exceptions import NoViewFoundError
-from dbally.iql_generator.iql_prompt_template import UnsupportedQueryError
+from dbally.iql_generator.prompt import UnsupportedQueryError
 from dbally.prompt import PromptTemplateError
